AD-621 Reactor Audit Batcher - Exposing a testing function in production code
